moved method fig_to_base64 to to utilities
