+++ protzilla/data_integration/di_plots.py
@@ -4,7 +4,7 @@ import pandas as pd
 from django.contrib import messages
 
 from ..constants.colors import PROTZILLA_DISCRETE_COLOR_SEQUENCE
-from protzilla.utilities.transform_figure import fig_to_base64
+from protzilla.utilities.utilities import fig_to_base64
 
 
 def go_enrichment_bar_plot(
