--- protzilla/utilities/transform_figure.py
@@ -1,17 +1,0 @@
-import base64
-import io
-
-
-def fig_to_base64(fig):
-    """
-    Convert a matplotlib figure to base64. This is used to display the figure in the browser.
-
-    :param fig: matplotlib figure
-    :type fig: matplotlib.figure.Figure
-    :return: base64 encoded image
-    :rtype: bytes
-    """
-    img = io.BytesIO()
-    fig.savefig(img, format="png", bbox_inches="tight")
-    img.seek(0)
-    return base64.b64encode(img.getvalue())
