+++ protzilla/utilities/utilities.py
@@ -3,7 +3,8 @@ import os
 from itertools import groupby
 from random import choices
 from string import ascii_letters
-
+import base64
+import io
 import psutil
 
 
@@ -32,3 +33,18 @@ def clean_uniprot_id(uniprot_id):
     if "_" in uniprot_id:
         uniprot_id = uniprot_id.split("_")[0]
     return uniprot_id
+
+
+def fig_to_base64(fig):
+    """
+    Convert a matplotlib figure to base64. This is used to display the figure in the browser.
+
+    :param fig: matplotlib figure
+    :type fig: matplotlib.figure.Figure
+    :return: base64 encoded image
+    :rtype: bytes
+    """
+    img = io.BytesIO()
+    fig.savefig(img, format="png", bbox_inches="tight")
+    img.seek(0)
+    return base64.b64encode(img.getvalue())
